requirements.txt
opsdroid
matrix-nio
pytest
aiohttp
pyjwt
httpx